move BitStringAddress code around, remove BStringAddress
